refactor(notice-details-page): modularized the code and optimized the performance
